# Optimize events, support objects with `handleEvent`