@tivac 2017-08-31
Putting some 👀 on perf. Changes look fine otherwise.