@isiahmeadows 2017-08-31
@tivac I didn't post my results here, but I did run the suite. It was a mild improvement in speed, but that wasn't the primary focus of the optimization (it's memory/GC).