@tivac 2017-08-31
Weird, I was seeing on the order of 1-5% **slower** perf. I got lost in the weeds some on making some tooling around comparing perf of branches easier though.