@isiahmeadows 2017-08-31
@tivac That's odd...I was seeing that in *gain*, so I guess we can assume it's within the margin of error. I'll note a pretty important detail: adding/removing event listeners is much more expensive in browsers than with our mock. Did you run the benchmarks in an actual browser or just Node? (I'd expect that if you're benching based on the mock, you *should* get a mild slowdown due to a slightly increased overhead.)

Another item of note: it no longer fast-paths things like `onclick` or `onsubmit`, where there are associated properties, but that's to simplify the combining logic and avoid creating closures where those APIs don't support the `EventHandler` interface.