@spacejack 2017-08-31
Using the `onclick` etc. properties was pretty crude in the way that it would [always set them to null](https://github.com/MithrilJS/mithril.js/blob/next/render/render.js#L586) when it found them on the element, whether or not a callback had previously been added. Though I don't know how expensive those read/writes are in practice.