@isiahmeadows 2017-08-31
@spacejack It's actually almost as cheap as normal property assignment in practice to modify the `on{event}` properties, since the native logic is so simple. With them, it's faster to just unconditionally overwrite them than to actually diff them. (With non-rendering attributes like `data-` attributes, it's generally pretty similar.)