@tivac 2017-09-01
Ah, yeah that explains it. If it's faster in browsers than I'm fine w/ this.