@isiahmeadows 2017-09-01
Okay...so here's some numbers, from benchmarking in Chrome:

Before this patch:

```
rerender identical vnode x 3,537,534 ops/sec ±22.62% (52 runs sampled)
rerender same tree x 38,454 ops/sec ±11.56% (40 runs sampled)
construct large VDOM tree x 24,846 ops/sec ±4.39% (48 runs sampled)
mutate styles/properties x 14,497 ops/sec ±6.56% (34 runs sampled)
repeated trees (recycling) x 325 ops/sec ±10.42% (43 runs sampled)
repeated trees (no recycling) x 336 ops/sec ±9.99% (45 runs sampled)
Completed perf tests in 95835ms
```

After this patch:

```
rerender identical vnode x 3,457,798 ops/sec ±30.39% (50 runs sampled)
rerender same tree x 77,396 ops/sec ±5.30% (44 runs sampled)
construct large VDOM tree x 24,041 ops/sec ±7.03% (43 runs sampled)
mutate styles/properties x 14,640 ops/sec ±5.48% (40 runs sampled)
repeated trees (recycling) x 329 ops/sec ±10.01% (39 runs sampled)
repeated trees (no recycling) x 336 ops/sec ±11.05% (42 runs sampled)
Completed perf tests in 94860ms
```

I'll come back with memory numbers later.