@isiahmeadows 2018-09-01
I added the "backport to v1" label. We can still keep most of the same code, just if it returns `false`, call `ev.preventDefault()` + `ev.stopPropagation()`. That's the only difference to retain compatibility.